Changed Bind.NotifyPropertyChanged method.
Changed it to use the Evaluator class to get the target object.
Now its no more an extension method to an object. Instead its a static
method that is called via Bind.NotifyPropertyChanged. My idea is that
the Bind class could function as a main point to trigger all property
change stuff.